change sum field from decimal to integer

+++ spending_app/forms/spending.py
@@ -1,12 +1,12 @@
 from spending_app.forms.baseForm import BaseForm
-from wtforms import DateField, DecimalField, StringField, IntegerField
+from wtforms import DateField, StringField, IntegerField
 from wtforms.validators import DataRequired
 
 
 class SpendingForm(BaseForm):
     id = IntegerField('Id')
     date = DateField('Дата', [DataRequired()])
-    sum = DecimalField('Сумма', [DataRequired()])
+    sum = IntegerField('Сумма', [DataRequired()])
     text = StringField('Описание')
     category_id = IntegerField('Категория', [DataRequired()])
